--- a/aggregate_test.go
+++ b/aggregate_test.go
@@ -433,6 +433,71 @@ func BenchmarkAggregation(b *testing.B) {
 	}
 }
 
+// Benchmark_ParquetAggregation is similar to BenchmarkAggregation but uses uses Parquet as the storage format.
+func Benchmark_ParquetAggregation(b *testing.B) {
+	ctx := context.Background()
+
+	columnStore, err := New()
+	require.NoError(b, err)
+	defer columnStore.Close()
+
+	db, err := columnStore.DB(ctx, "test")
+	require.NoError(b, err)
+
+	// Insert sample data
+	{
+		config := NewTableConfig(dynparquet.SampleDefinition())
+		table, err := db.Table("test", config)
+		require.NoError(b, err)
+
+		samples := make(dynparquet.Samples, 0, 200_000)
+		for i := 0; i < cap(samples); i++ {
+			samples = append(samples, dynparquet.Sample{
+				Labels: map[string]string{
+					"label1": "value1",
+					"label2": "value" + strconv.Itoa(i%3),
+				},
+				Stacktrace: []uuid.UUID{
+					{0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x1},
+				},
+				Timestamp: int64(i),
+				Value:     int64(i),
+			})
+		}
+
+		r, err := samples.ToRecord()
+		require.NoError(b, err)
+		_, err = table.InsertRecord(ctx, r)
+		require.NoError(b, err)
+		require.NoError(b, table.EnsureCompaction())
+	}
+
+	engine := query.NewEngine(
+		memory.NewGoAllocator(),
+		db.TableProvider(),
+	)
+
+	b.ResetTimer()
+	b.Run("paruqet_aggregation", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			engine.ScanTable("test").
+				Filter(
+					logicalplan.Col("timestamp").Gt(logicalplan.Literal(10)),
+				).
+				Aggregate(
+					[]*logicalplan.AggregationFunction{
+						logicalplan.Sum(logicalplan.Col("value")),
+					},
+					[]logicalplan.Expr{
+						logicalplan.Col("labels.label2"),
+					},
+				).Execute(ctx, func(ctx context.Context, r arrow.Record) error {
+				return nil
+			})
+		}
+	})
+}
+
 func Test_Aggregation_DynCol(t *testing.T) {
 	mem := memory.NewCheckedAllocator(memory.DefaultAllocator)
 	defer mem.AssertSize(t, 0)