@thorfour 2024-05-22
Ooof I found that there was a bug in this change where if we had more than two pages the offset i was incorrect and we started to overwrite pages. The different when that's fixed is minimal. Going to close